Move the syntax of each kind into a helper of its own
`IntroduceNamedTypeTransformation` carried the syntax of every kind beside the switch
that chooses between them, which made the file grow with each kind added.

`EnumHelper`, `DelegateHelper`, `RecordHelper` and `UnionHelper` carry what is specific
to one kind, and the transformation keeps the switch and what every kind shares. The
union helper is compiled into the latest Roslyn variant only, so the condition that the
transformation carried for it is now the condition of a whole file.
